// ui/src/orval/models/navigationTreeDatabase.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * UI Router API
 * Defines the specification for the UI Catalog API
 * OpenAPI spec version: 1.0.2
 */
import type { NavigationTreeSchema } from './navigationTreeSchema';

export interface NavigationTreeDatabase {
  name: string;
  schemas: NavigationTreeSchema[];
}